Cleaned up Today's work on Client and Credentials
Still not working properly, finding it difficult to debug with
async/await.  Will take a look tomorrow or Monday.  Can now load from
text files, and the Login methods are working in the chat client.
Again, a basic port of Echo Ninja, but still just trying to get things
online.